+++ image-viewer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Image Viewer Changelog
 
+## 1.0.0 - 2018-07-18
+## Change
+- Image links detection redone
+- Update expand/shrink icons
+
+## Added
+- Image host support: `radikal.ru`, `stuffed.ru`, `imageban.ru`, `piccash.net`, `imgdrive.net`, `imgchilibum.ru`, `XXXScreens.com`, `money-pic.ru`
+- Display host alias urls on hover
+
+## Fixed
+- Backdrop isn't hidden after image view close
+
 ## 0.9.0 - 2018-04-22
 ## Added
 - Separate from ple